requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sentence-transformers==2.3.1
scikit-learn==1.4.0
textstat==0.7.3
openai==1.10.0
httpx==0.26.0
streamlit==1.30.0
plotly==5.18.0
python-multipart==0.0.6
google-genai==1.60.0
python-dotenv==1.0.1
